+++ x/bank/keeper/send.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"context"
 	"fmt"
 
 	gogotypes "github.com/cosmos/gogoproto/types"
@@ -19,6 +20,10 @@ import (
 // between accounts without the possibility of creating coins.
 type SendKeeper interface {
 	ViewKeeper
+
+	AppendSendRestriction(restriction types.SendRestrictionFn)
+	PrependSendRestriction(restriction types.SendRestrictionFn)
+	ClearSendRestriction()
 
 	InputOutputCoins(ctx sdk.Context, inputs []types.Input, outputs []types.Output) error
 	SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) error
@@ -60,6 +65,8 @@ type BaseSendKeeper struct {
 	// the address capable of executing a MsgUpdateParams message. Typically, this
 	// should be the x/gov module account.
 	authority string
+
+	sendRestriction *sendRestriction
 }
 
 func NewBaseSendKeeper(
@@ -75,13 +82,29 @@ func NewBaseSendKeeper(
 	}
 
 	return BaseSendKeeper{
-		BaseViewKeeper: NewBaseViewKeeper(cdc, storeKey, tStoreKey, ak),
-		cdc:            cdc,
-		ak:             ak,
-		storeKey:       storeKey,
-		blockedAddrs:   blockedAddrs,
-		authority:      authority,
+		BaseViewKeeper:  NewBaseViewKeeper(cdc, storeKey, tStoreKey, ak),
+		cdc:             cdc,
+		ak:              ak,
+		storeKey:        storeKey,
+		blockedAddrs:    blockedAddrs,
+		authority:       authority,
+		sendRestriction: newSendRestriction(),
 	}
+}
+
+// AppendSendRestriction adds the provided SendRestrictionFn to run after previously provided restrictions.
+func (k BaseSendKeeper) AppendSendRestriction(restriction types.SendRestrictionFn) {
+	k.sendRestriction.append(restriction)
+}
+
+// PrependSendRestriction adds the provided SendRestrictionFn to run before previously provided restrictions.
+func (k BaseSendKeeper) PrependSendRestriction(restriction types.SendRestrictionFn) {
+	k.sendRestriction.prepend(restriction)
+}
+
+// ClearSendRestriction removes the send restriction (if there is one).
+func (k BaseSendKeeper) ClearSendRestriction() {
+	k.sendRestriction.clear()
 }
 
 // GetAuthority returns the x/bank module's authority.
@@ -162,6 +185,11 @@ func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, inputs []types.Input, 
 			return err
 		}
 
+		outAddress, err = k.sendRestriction.apply(ctx, inAddress, outAddress, out.Coins)
+		if err != nil {
+			return err
+		}
+
 		if err := k.addCoins(ctx, outAddress, out.Coins); err != nil {
 			return err
 		}
@@ -169,7 +197,7 @@ func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, inputs []types.Input, 
 		ctx.EventManager().EmitEvent(
 			sdk.NewEvent(
 				types.EventTypeTransfer,
-				sdk.NewAttribute(types.AttributeKeyRecipient, out.Address),
+				sdk.NewAttribute(types.AttributeKeyRecipient, outAddress.String()),
 				sdk.NewAttribute(sdk.AttributeKeyAmount, out.Coins.String()),
 			),
 		)
@@ -192,6 +220,11 @@ func (k BaseSendKeeper) InputOutputCoins(ctx sdk.Context, inputs []types.Input, 
 // An error is returned upon failure.
 func (k BaseSendKeeper) SendCoins(ctx sdk.Context, fromAddr sdk.AccAddress, toAddr sdk.AccAddress, amt sdk.Coins) error {
 	err := k.subUnlockedCoins(ctx, fromAddr, amt)
+	if err != nil {
+		return err
+	}
+
+	toAddr, err = k.sendRestriction.apply(ctx, fromAddr, toAddr, amt)
 	if err != nil {
 		return err
 	}
@@ -523,4 +556,42 @@ func (k BaseSendKeeper) getSendEnabledOrDefault(store sdk.KVStore, denom string,
 	}
 
 	return defaultVal
+}
+
+// sendRestriction is a struct that houses a SendRestrictionFn.
+// It exists so that the SendRestrictionFn can be updated in the SendKeeper without needing to have a pointer receiver.
+type sendRestriction struct {
+	fn types.SendRestrictionFn
+}
+
+// newSendRestriction creates a new sendRestriction with nil send restriction.
+func newSendRestriction() *sendRestriction {
+	return &sendRestriction{
+		fn: nil,
+	}
+}
+
+// append adds the provided restriction to this, to be run after the existing function.
+func (r *sendRestriction) append(restriction types.SendRestrictionFn) {
+	r.fn = r.fn.Then(restriction)
+}
+
+// prepend adds the provided restriction to this, to be run before the existing function.
+func (r *sendRestriction) prepend(restriction types.SendRestrictionFn) {
+	r.fn = restriction.Then(r.fn)
+}
+
+// clear removes the send restriction (sets it to nil).
+func (r *sendRestriction) clear() {
+	r.fn = nil
+}
+
+var _ types.SendRestrictionFn = (*sendRestriction)(nil).apply
+
+// apply applies the send restriction if there is one. If not, it's a no-op.
+func (r *sendRestriction) apply(ctx context.Context, fromAddr, toAddr sdk.AccAddress, amt sdk.Coins) (sdk.AccAddress, error) {
+	if r == nil || r.fn == nil {
+		return toAddr, nil
+	}
+	return r.fn(ctx, fromAddr, toAddr, amt)
 }